+++ kinetic/client.py
@@ -95,6 +95,7 @@ class Client(BaseClient):
     def mediaOptimize(self, *args, **kwargs):
         return self._process(operations.MediaOptimize, *args, **kwargs)
 
+
 class KineticRangeIter(object):
 
     def __init__(self, client, startKey, endKey, startKeyInclusive,endKeyInclusive, prefetch):